// src/config/interfaces/interfaces.ts
export interface IWebStory {
  id: number;
  title: string;
  meta: {
    first_published_at: string;
  };
  story_thumbnail: {
    meta: {
      download_url: string;
    };
  };
  subcategories: ICategory[];
  categories: ICategory[];
  webstory_slug: string;
}
// ...
export interface ICategory {
  id: number;
  name: string;
  slug: string;
  meta_title: string;
  watermark: string;
  meta_discriptions: string;
}